requirements.txt
# Avatar Tutor — dependențe Python
# Instalare: pip install -r requirements.txt

# ── GUI ───────────────────────────────────────────────
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# ── Computer Vision & Procesare imagini ───────────────
opencv-python>=4.9.0
numpy>=1.26.0
mediapipe>=0.10.0
Pillow>=10.0.0
matplotlib>=3.8.0

# ── Audio & Voice ─────────────────────────────────────
sounddevice>=0.4.6
faster-whisper>=1.0.0
scipy>=1.11.0

# ── TTS (Text-to-Speech) ──────────────────────────────
elevenlabs>=1.0.0
# piper-tts: instalare separată — vezi README.md

# ── LLM / AI APIs ─────────────────────────────────────
google-genai>=0.3.0
requests>=2.31.0
aiohttp>=3.9.0

# ── Utilități ─────────────────────────────────────────
python-dotenv>=1.0.0
openpyxl>=3.1.0
